refactor: rename taskTimelineTypeRegistry to messageColors and extract shouldShowInTimeline
- Rename taskTimelineTypeRegistry.ts to messageColors.ts for better semantic clarity
- Rename taskTimelineColorsByType to taskTimelineColorPalette for consistency
- Extract shouldShowInTimeline function to messageColors module for reusability
- Update all imports to use the new module name
- Consolidate message color logic in a single, well-named module
