+++ webview-ui/src/components/chat/TaskTimelineMessage.tsx
@@ -2,7 +2,7 @@ import { memo, useEffect, useState } from "react"
 import { useTranslation } from "react-i18next"
 import type { ClineMessage } from "@roo-code/types"
 import { cn } from "@/lib/utils"
-import { getMessageTypeDescription } from "@/utils/timeline/taskTimelineTypeRegistry"
+import { getMessageTypeDescription } from "@/utils/messageColors"
 import { MAX_HEIGHT_PX, type MessageSizeData } from "@/utils/timeline/calculateTaskTimelineSizes"
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip"
