+++ webview-ui/src/utils/messageColors.ts
@@ -5,7 +5,7 @@ export interface TaskTimelineMessageTypeConfig {
 	translationKey: string
 }
 
-const taskTimelineColorsByType = {
+export const taskTimelineColorPalette = {
 	USER_INTERACTION: "bg-[var(--vscode-chat-editedFileForeground)]", // Tan/orange for all 'ask' types
 	SYSTEM_READ: "bg-[var(--vscode-textLink-foreground)]", // Light blue for file reads
 	SYSTEM_WRITE: "bg-[var(--vscode-focusBorder)]", // Dark blue for file writes
@@ -25,81 +25,81 @@ const taskTimelineColorsByType = {
 export const TASK_TIMELINE_MESSAGE_TYPES: Record<string, TaskTimelineMessageTypeConfig> = {
 	// Ask types that should be shown (everything except the filtered ones)
 	"ask:tool": {
-		color: taskTimelineColorsByType.SYSTEM_GENERAL_TOOL,
+		color: taskTimelineColorPalette.SYSTEM_GENERAL_TOOL,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.tool",
 	},
 	"ask:completion_result": {
-		color: taskTimelineColorsByType.SUCCESS,
+		color: taskTimelineColorPalette.SUCCESS,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.completion_result",
 	},
 	"ask:browser_action_launch": {
-		color: taskTimelineColorsByType.SYSTEM_GENERAL_TOOL,
+		color: taskTimelineColorPalette.SYSTEM_GENERAL_TOOL,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.browser_action_launch",
 	},
 	"ask:use_mcp_server": {
-		color: taskTimelineColorsByType.SYSTEM_GENERAL_TOOL,
+		color: taskTimelineColorPalette.SYSTEM_GENERAL_TOOL,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.use_mcp_server",
 	},
 	"ask:followup": {
-		color: taskTimelineColorsByType.USER_INTERACTION,
+		color: taskTimelineColorPalette.USER_INTERACTION,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.followup",
 	},
 	"ask:command": {
-		color: taskTimelineColorsByType.USER_INTERACTION,
+		color: taskTimelineColorPalette.USER_INTERACTION,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.command",
 	},
 
 	// Say types that should be shown (everything except the filtered ones)
 	"say:text": {
-		color: taskTimelineColorsByType.ASSISTANT_MUTTERING,
+		color: taskTimelineColorPalette.ASSISTANT_MUTTERING,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.text",
 	},
 	"say:reasoning": {
-		color: taskTimelineColorsByType.ASSISTANT_MUTTERING,
+		color: taskTimelineColorPalette.ASSISTANT_MUTTERING,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.reasoning",
 	},
 	"say:mcp_server_response": {
-		color: taskTimelineColorsByType.SYSTEM_GENERAL_TOOL,
+		color: taskTimelineColorPalette.SYSTEM_GENERAL_TOOL,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.mcp_server_response",
 	},
 	"say:command_output": {
-		color: taskTimelineColorsByType.SYSTEM_GENERAL_TOOL,
+		color: taskTimelineColorPalette.SYSTEM_GENERAL_TOOL,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.command_output",
 	},
 	"say:browser_action": {
-		color: taskTimelineColorsByType.SYSTEM_GENERAL_TOOL,
+		color: taskTimelineColorPalette.SYSTEM_GENERAL_TOOL,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.browser_action",
 	},
 	"say:browser_action_result": {
-		color: taskTimelineColorsByType.SYSTEM_GENERAL_TOOL,
+		color: taskTimelineColorPalette.SYSTEM_GENERAL_TOOL,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.browser_action_result",
 	},
 	"say:completion_result": {
-		color: taskTimelineColorsByType.SUCCESS,
+		color: taskTimelineColorPalette.SUCCESS,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.completion_result",
 	},
 	"say:error": {
-		color: taskTimelineColorsByType.ERROR,
+		color: taskTimelineColorPalette.ERROR,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.error",
 	},
 	"say:checkpoint_saved": {
-		color: taskTimelineColorsByType.SUCCESS,
+		color: taskTimelineColorPalette.SUCCESS,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.checkpoint_saved",
 	},
 	"say:condense_context": {
-		color: taskTimelineColorsByType.ASSISTANT_MUTTERING,
+		color: taskTimelineColorPalette.ASSISTANT_MUTTERING,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.condense_context",
 	},
 	"say:user_feedback": {
-		color: taskTimelineColorsByType.USER_INTERACTION,
+		color: taskTimelineColorPalette.USER_INTERACTION,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.user",
 	},
 	"say:user_feedback_diff": {
-		color: taskTimelineColorsByType.USER_INTERACTION,
+		color: taskTimelineColorPalette.USER_INTERACTION,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.user",
 	},
 	"ask:condense": {
-		color: taskTimelineColorsByType.ASSISTANT_MUTTERING,
+		color: taskTimelineColorPalette.ASSISTANT_MUTTERING,
 		translationKey: "kilocode:taskTimeline.tooltip.messageTypes.condense",
 	},
 }
@@ -162,7 +162,7 @@ export function isFileOperation(message: ClineMessage): { isFile: boolean; isWri
  */
 export function getTaskTimelineMessageColor(message: ClineMessage | ClineMessage[]): string {
 	if (Array.isArray(message)) {
-		return taskTimelineColorsByType.GROUPED
+		return taskTimelineColorPalette.GROUPED
 	}
 
 	const singleMessage = message as ClineMessage
@@ -172,13 +172,16 @@ export function getTaskTimelineMessageColor(message: ClineMessage | ClineMessage
 	if (singleMessage.type === "ask") {
 		const fileOp = isFileOperation(singleMessage)
 		if (fileOp.isFile) {
-			return fileOp.isWrite ? taskTimelineColorsByType.SYSTEM_WRITE : taskTimelineColorsByType.SYSTEM_READ
+			return fileOp.isWrite ? taskTimelineColorPalette.SYSTEM_WRITE : taskTimelineColorPalette.SYSTEM_READ
 		}
 	}
 
 	// Use registry color or fallback to default
 	const config = TASK_TIMELINE_MESSAGE_TYPES[messageKey]
-	return config ? config.color : taskTimelineColorsByType.DEFAULT
+	return config ? config.color : taskTimelineColorPalette.DEFAULT
 }
 
-export { taskTimelineColorsByType as taskTimelineColorPalette }
+export function shouldShowInTimeline(message: ClineMessage): boolean {
+	const messageKey = getTaskTimelineMessageTypeKey(message)
+	return messageKey in TASK_TIMELINE_MESSAGE_TYPES
+}
